fallback for RelyingParty
Updated the constant variables located at src\constants.js that were using https://bedrock.frontendlegacy.realms.minecraft-services.net as the RelyingParty.

The issue occured when using this endpoint as the RelyingParty when authenticating to Minecraft Services, the request will fail. I'm not sure if this was strictly a post 1.21.130 thing but that's when the issue was noticed.

Using the old https://pocket.realms.minecraft.net/ instead has produced no errors.